+++ cntrlr/src/hw/mcu/kinetis/peripheral/port.rs
@@ -8,7 +8,7 @@
 
 use super::{
     super::{Mk20Dx128, Mk20Dx256, Mk64Fx512, Mk66Fx1M0, Mkl26Z64},
-    sim::{Gate, Peripheral},
+    sim::{Gate, GatedPeripheral},
 };
 use crate::{digital::Pull, register::Register, sync::Flag};
 use bit_field::BitField;
@@ -590,7 +590,7 @@ impl super::uart::UartTx<Mkl26Z64, 0> for UartTx<Pin<'_, Mkl26Z64, 1, 17>> {}
 impl super::uart::UartTx<Mkl26Z64, 1> for UartTx<Pin<'_, Mkl26Z64, 2, 4>> {}
 impl super::uart::UartTx<Mkl26Z64, 2> for UartTx<Pin<'_, Mkl26Z64, 3, 3>> {}
 
-unsafe impl Peripheral<Mk20Dx128> for Port<Mk20Dx128, 0> {
+unsafe impl GatedPeripheral<Mk20Dx128> for Port<Mk20Dx128, 0> {
     const GATE: (usize, usize) = (5, 9);
 
     unsafe fn new(gate: Gate) -> Self {
@@ -603,7 +603,7 @@ unsafe impl Peripheral<Mk20Dx128> for Port<Mk20Dx128, 0> {
     }
 }
 
-unsafe impl Peripheral<Mk20Dx128> for Port<Mk20Dx128, 1> {
+unsafe impl GatedPeripheral<Mk20Dx128> for Port<Mk20Dx128, 1> {
     const GATE: (usize, usize) = (5, 10);
 
     unsafe fn new(gate: Gate) -> Self {
@@ -616,7 +616,7 @@ unsafe impl Peripheral<Mk20Dx128> for Port<Mk20Dx128, 1> {
     }
 }
 
-unsafe impl Peripheral<Mk20Dx128> for Port<Mk20Dx128, 2> {
+unsafe impl GatedPeripheral<Mk20Dx128> for Port<Mk20Dx128, 2> {
     const GATE: (usize, usize) = (5, 11);
 
     unsafe fn new(gate: Gate) -> Self {
@@ -629,7 +629,7 @@ unsafe impl Peripheral<Mk20Dx128> for Port<Mk20Dx128, 2> {
     }
 }
 
-unsafe impl Peripheral<Mk20Dx128> for Port<Mk20Dx128, 3> {
+unsafe impl GatedPeripheral<Mk20Dx128> for Port<Mk20Dx128, 3> {
     const GATE: (usize, usize) = (5, 12);
 
     unsafe fn new(gate: Gate) -> Self {
@@ -642,7 +642,7 @@ unsafe impl Peripheral<Mk20Dx128> for Port<Mk20Dx128, 3> {
     }
 }
 
-unsafe impl Peripheral<Mk20Dx128> for Port<Mk20Dx128, 4> {
+unsafe impl GatedPeripheral<Mk20Dx128> for Port<Mk20Dx128, 4> {
     const GATE: (usize, usize) = (5, 13);
 
     unsafe fn new(gate: Gate) -> Self {
@@ -655,7 +655,7 @@ unsafe impl Peripheral<Mk20Dx128> for Port<Mk20Dx128, 4> {
     }
 }
 
-unsafe impl Peripheral<Mk20Dx256> for Port<Mk20Dx256, 0> {
+unsafe impl GatedPeripheral<Mk20Dx256> for Port<Mk20Dx256, 0> {
     const GATE: (usize, usize) = (5, 9);
 
     unsafe fn new(gate: Gate) -> Self {
@@ -668,7 +668,7 @@ unsafe impl Peripheral<Mk20Dx256> for Port<Mk20Dx256, 0> {
     }
 }
 
-unsafe impl Peripheral<Mk20Dx256> for Port<Mk20Dx256, 1> {
+unsafe impl GatedPeripheral<Mk20Dx256> for Port<Mk20Dx256, 1> {
     const GATE: (usize, usize) = (5, 10);
 
     unsafe fn new(gate: Gate) -> Self {
@@ -681,7 +681,7 @@ unsafe impl Peripheral<Mk20Dx256> for Port<Mk20Dx256, 1> {
     }
 }
 
-unsafe impl Peripheral<Mk20Dx256> for Port<Mk20Dx256, 2> {
+unsafe impl GatedPeripheral<Mk20Dx256> for Port<Mk20Dx256, 2> {
     const GATE: (usize, usize) = (5, 11);
 
     unsafe fn new(gate: Gate) -> Self {
@@ -694,7 +694,7 @@ unsafe impl Peripheral<Mk20Dx256> for Port<Mk20Dx256, 2> {
     }
 }
 
-unsafe impl Peripheral<Mk20Dx256> for Port<Mk20Dx256, 3> {
+unsafe impl GatedPeripheral<Mk20Dx256> for Port<Mk20Dx256, 3> {
     const GATE: (usize, usize) = (5, 12);
 
     unsafe fn new(gate: Gate) -> Self {
@@ -707,7 +707,7 @@ unsafe impl Peripheral<Mk20Dx256> for Port<Mk20Dx256, 3> {
     }
 }
 
-unsafe impl Peripheral<Mk20Dx256> for Port<Mk20Dx256, 4> {
+unsafe impl GatedPeripheral<Mk20Dx256> for Port<Mk20Dx256, 4> {
     const GATE: (usize, usize) = (5, 13);
 
     unsafe fn new(gate: Gate) -> Self {
@@ -720,7 +720,7 @@ unsafe impl Peripheral<Mk20Dx256> for Port<Mk20Dx256, 4> {
     }
 }
 
-unsafe impl Peripheral<Mk64Fx512> for Port<Mk64Fx512, 0> {
+unsafe impl GatedPeripheral<Mk64Fx512> for Port<Mk64Fx512, 0> {
     const GATE: (usize, usize) = (5, 9);
 
     unsafe fn new(gate: Gate) -> Self {
@@ -733,7 +733,7 @@ unsafe impl Peripheral<Mk64Fx512> for Port<Mk64Fx512, 0> {
     }
 }
 
-unsafe impl Peripheral<Mk64Fx512> for Port<Mk64Fx512, 1> {
+unsafe impl GatedPeripheral<Mk64Fx512> for Port<Mk64Fx512, 1> {
     const GATE: (usize, usize) = (5, 10);
 
     unsafe fn new(gate: Gate) -> Self {
@@ -746,7 +746,7 @@ unsafe impl Peripheral<Mk64Fx512> for Port<Mk64Fx512, 1> {
     }
 }
 
-unsafe impl Peripheral<Mk64Fx512> for Port<Mk64Fx512, 2> {
+unsafe impl GatedPeripheral<Mk64Fx512> for Port<Mk64Fx512, 2> {
     const GATE: (usize, usize) = (5, 11);
 
     unsafe fn new(gate: Gate) -> Self {
@@ -759,7 +759,7 @@ unsafe impl Peripheral<Mk64Fx512> for Port<Mk64Fx512, 2> {
     }
 }
 
-unsafe impl Peripheral<Mk64Fx512> for Port<Mk64Fx512, 3> {
+unsafe impl GatedPeripheral<Mk64Fx512> for Port<Mk64Fx512, 3> {
     const GATE: (usize, usize) = (5, 12);
 
     unsafe fn new(gate: Gate) -> Self {
@@ -772,7 +772,7 @@ unsafe impl Peripheral<Mk64Fx512> for Port<Mk64Fx512, 3> {
     }
 }
 
-unsafe impl Peripheral<Mk64Fx512> for Port<Mk64Fx512, 4> {
+unsafe impl GatedPeripheral<Mk64Fx512> for Port<Mk64Fx512, 4> {
     const GATE: (usize, usize) = (5, 13);
 
     unsafe fn new(gate: Gate) -> Self {
@@ -785,7 +785,7 @@ unsafe impl Peripheral<Mk64Fx512> for Port<Mk64Fx512, 4> {
     }
 }
 
-unsafe impl Peripheral<Mk66Fx1M0> for Port<Mk66Fx1M0, 0> {
+unsafe impl GatedPeripheral<Mk66Fx1M0> for Port<Mk66Fx1M0, 0> {
     const GATE: (usize, usize) = (5, 9);
 
     unsafe fn new(gate: Gate) -> Self {
@@ -798,7 +798,7 @@ unsafe impl Peripheral<Mk66Fx1M0> for Port<Mk66Fx1M0, 0> {
     }
 }
 
-unsafe impl Peripheral<Mk66Fx1M0> for Port<Mk66Fx1M0, 1> {
+unsafe impl GatedPeripheral<Mk66Fx1M0> for Port<Mk66Fx1M0, 1> {
     const GATE: (usize, usize) = (5, 10);
 
     unsafe fn new(gate: Gate) -> Self {
@@ -811,7 +811,7 @@ unsafe impl Peripheral<Mk66Fx1M0> for Port<Mk66Fx1M0, 1> {
     }
 }
 
-unsafe impl Peripheral<Mk66Fx1M0> for Port<Mk66Fx1M0, 2> {
+unsafe impl GatedPeripheral<Mk66Fx1M0> for Port<Mk66Fx1M0, 2> {
     const GATE: (usize, usize) = (5, 11);
 
     unsafe fn new(gate: Gate) -> Self {
@@ -824,7 +824,7 @@ unsafe impl Peripheral<Mk66Fx1M0> for Port<Mk66Fx1M0, 2> {
     }
 }
 
-unsafe impl Peripheral<Mk66Fx1M0> for Port<Mk66Fx1M0, 3> {
+unsafe impl GatedPeripheral<Mk66Fx1M0> for Port<Mk66Fx1M0, 3> {
     const GATE: (usize, usize) = (5, 12);
 
     unsafe fn new(gate: Gate) -> Self {
@@ -837,7 +837,7 @@ unsafe impl Peripheral<Mk66Fx1M0> for Port<Mk66Fx1M0, 3> {
     }
 }
 
-unsafe impl Peripheral<Mk66Fx1M0> for Port<Mk66Fx1M0, 4> {
+unsafe impl GatedPeripheral<Mk66Fx1M0> for Port<Mk66Fx1M0, 4> {
     const GATE: (usize, usize) = (5, 13);
 
     unsafe fn new(gate: Gate) -> Self {
@@ -850,7 +850,7 @@ unsafe impl Peripheral<Mk66Fx1M0> for Port<Mk66Fx1M0, 4> {
     }
 }
 
-unsafe impl Peripheral<Mkl26Z64> for Port<Mkl26Z64, 0> {
+unsafe impl GatedPeripheral<Mkl26Z64> for Port<Mkl26Z64, 0> {
     const GATE: (usize, usize) = (5, 9);
 
     unsafe fn new(gate: Gate) -> Self {
@@ -863,7 +863,7 @@ unsafe impl Peripheral<Mkl26Z64> for Port<Mkl26Z64, 0> {
     }
 }
 
-unsafe impl Peripheral<Mkl26Z64> for Port<Mkl26Z64, 1> {
+unsafe impl GatedPeripheral<Mkl26Z64> for Port<Mkl26Z64, 1> {
     const GATE: (usize, usize) = (5, 10);
 
     unsafe fn new(gate: Gate) -> Self {
@@ -876,7 +876,7 @@ unsafe impl Peripheral<Mkl26Z64> for Port<Mkl26Z64, 1> {
     }
 }
 
-unsafe impl Peripheral<Mkl26Z64> for Port<Mkl26Z64, 2> {
+unsafe impl GatedPeripheral<Mkl26Z64> for Port<Mkl26Z64, 2> {
     const GATE: (usize, usize) = (5, 11);
 
     unsafe fn new(gate: Gate) -> Self {
@@ -889,7 +889,7 @@ unsafe impl Peripheral<Mkl26Z64> for Port<Mkl26Z64, 2> {
     }
 }
 
-unsafe impl Peripheral<Mkl26Z64> for Port<Mkl26Z64, 3> {
+unsafe impl GatedPeripheral<Mkl26Z64> for Port<Mkl26Z64, 3> {
     const GATE: (usize, usize) = (5, 12);
 
     unsafe fn new(gate: Gate) -> Self {
@@ -902,7 +902,7 @@ unsafe impl Peripheral<Mkl26Z64> for Port<Mkl26Z64, 3> {
     }
 }
 
-unsafe impl Peripheral<Mkl26Z64> for Port<Mkl26Z64, 4> {
+unsafe impl GatedPeripheral<Mkl26Z64> for Port<Mkl26Z64, 4> {
     const GATE: (usize, usize) = (5, 13);
 
     unsafe fn new(gate: Gate) -> Self {